# rework byval ZIR instructions; forbid runtime vector indexes