--- a/lib/std/json/static.zig
+++ b/lib/std/json/static.zig
@@ -389,7 +389,7 @@ pub fn innerParse(
             switch (try source.peekNextTokenType()) {
                 .array_begin => {
                     // Typical array.
-                    return internalParseArray(T, arrayInfo.child, arrayInfo.len, allocator, source, options);
+                    return internalParseArray(T, arrayInfo.child, allocator, source, options);
                 },
                 .string => {
                     if (arrayInfo.child != u8) return error.UnexpectedToken;
@@ -443,7 +443,7 @@ pub fn innerParse(
         .vector => |vecInfo| {
             switch (try source.peekNextTokenType()) {
                 .array_begin => {
-                    return internalParseArray(T, vecInfo.child, vecInfo.len, allocator, source, options);
+                    return internalParseVector(T, vecInfo.child, vecInfo.len, allocator, source, options);
                 },
                 else => return error.UnexpectedToken,
             }
@@ -519,6 +519,25 @@ pub fn innerParse(
 fn internalParseArray(
     comptime T: type,
     comptime Child: type,
+    allocator: Allocator,
+    source: anytype,
+    options: ParseOptions,
+) !T {
+    assert(.array_begin == try source.next());
+
+    var r: T = undefined;
+    for (&r) |*elem| {
+        elem.* = try innerParse(Child, allocator, source, options);
+    }
+
+    if (.array_end != try source.next()) return error.UnexpectedToken;
+
+    return r;
+}
+
+fn internalParseVector(
+    comptime T: type,
+    comptime Child: type,
     comptime len: comptime_int,
     allocator: Allocator,
     source: anytype,
@@ -527,8 +546,7 @@ fn internalParseArray(
     assert(.array_begin == try source.next());
 
     var r: T = undefined;
-    var i: usize = 0;
-    while (i < len) : (i += 1) {
+    inline for (0..len) |i| {
         r[i] = try innerParse(Child, allocator, source, options);
     }