fixes #22906
fixes #13938
fixes #25111
partially addresses #24061 by making vector indexes comptime-known
partially addresses #5973

## Upgrade Guide

```zig
for (0..vector_len) |i| {
   _ = vector[i];
}
```

:arrow_down: 

```zig
// coerce the vector to an array
const vector_type = @typeInfo(@TypeOf(vector)).vector;
const array: [vector_type.len]vector_type.child = vector;
for (&array) |elem| {
    _ = elem;
}
```

## Performance Data Points

### llvm backend ReleaseFast compiler building Debug Zig compiler, master vs branch

```
Benchmark 1 (3 runs): master/bin/zig build-exe ...
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          12.3s  ± 16.2ms    12.3s  … 12.3s           0 ( 0%)        0%
  peak_rss           1.36GB ± 22.5MB    1.34GB … 1.39GB          0 ( 0%)        0%
  cpu_cycles          105G  ±  291M      105G  …  105G           0 ( 0%)        0%
  instructions        212G  ± 3.89M      212G  …  212G           0 ( 0%)        0%
  cache_references   6.67G  ± 19.6M     6.66G  … 6.69G           0 ( 0%)        0%
  cache_misses        466M  ± 5.03M      461M  …  471M           0 ( 0%)        0%
  branch_misses       404M  ± 1.57M      403M  …  406M           0 ( 0%)        0%
Benchmark 2 (3 runs): branch/bin/zig build-exe ...
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          12.9s  ± 36.4ms    12.9s  … 13.0s           0 ( 0%)        💩+  5.1% ±  0.5%
  peak_rss           1.40GB ± 8.90MB    1.40GB … 1.41GB          0 ( 0%)          +  3.0% ±  2.8%
  cpu_cycles          109G  ±  170M      109G  …  109G           0 ( 0%)        💩+  4.2% ±  0.5%
  instructions        221G  ± 5.81M      221G  …  221G           0 ( 0%)        💩+  4.0% ±  0.0%
  cache_references   7.08G  ± 8.88M     7.07G  … 7.09G           0 ( 0%)        💩+  6.1% ±  0.5%
  cache_misses        486M  ± 3.12M      483M  …  489M           0 ( 0%)        💩+  4.4% ±  2.0%
  branch_misses       413M  ± 1.25M      412M  …  414M           0 ( 0%)        💩+  2.1% ±  0.8%
```

### Zig Compiler ReleaseSmall Size

:hankey: 13.6 -> 14.1M (+4%) 

### The test case from #13938

* Debug :zap:  29677949 -> 236675ns (-99%)
* ReleaseFast :zap: 28415587 -> 491ns (-100%)

###  llvm backend ReleaseFast building Debug Zig compiler, using llvm backend vs using x86 backend, with the branch

```
Benchmark 1 (3 runs): stage3/bin/zig build-exe -fllvm -flld ...
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          60.0s  ± 63.6ms    59.9s  … 60.0s           0 ( 0%)        0%
  peak_rss           4.01GB ±  447KB    4.01GB … 4.01GB          0 ( 0%)        0%
  cpu_cycles          316G  ±  385M      315G  …  316G           0 ( 0%)        0%
  instructions        583G  ±  221M      583G  …  583G           0 ( 0%)        0%
  cache_references   22.4G  ± 49.5M     22.3G  … 22.4G           0 ( 0%)        0%
  cache_misses       4.54G  ± 9.15M     4.53G  … 4.55G           0 ( 0%)        0%
  branch_misses      2.27G  ± 1.79M     2.27G  … 2.27G           0 ( 0%)        0%
Benchmark 2 (3 runs): stage3/bin/zig build-exe -fno-llvm -fno-lld ...
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          13.0s  ± 29.7ms    13.0s  … 13.1s           0 ( 0%)        ⚡- 78.3% ±  0.2%
  peak_rss           1.38GB ± 31.0MB    1.34GB … 1.41GB          0 ( 0%)        ⚡- 65.7% ±  1.2%
  cpu_cycles          111G  ±  136M      110G  …  111G           0 ( 0%)        ⚡- 65.0% ±  0.2%
  instructions        222G  ± 10.5M      222G  …  222G           0 ( 0%)        ⚡- 62.0% ±  0.1%
  cache_references   7.17G  ± 10.5M     7.16G  … 7.18G           0 ( 0%)        ⚡- 67.9% ±  0.4%
  cache_misses        482M  ± 3.52M      478M  …  485M           0 ( 0%)        ⚡- 89.4% ±  0.3%
  branch_misses       424M  ± 1.23M      423M  …  425M           0 ( 0%)        ⚡- 81.3% ±  0.2%
```

### x86 backend Debug compiler building Debug Zig compiler, master vs branch

```
Benchmark 1 (3 runs): master/stage4/bin/zig build-exe ...
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           423s  ± 4.75s      420s  …  428s           0 ( 0%)        0%
  peak_rss           3.59GB ±  166MB    3.50GB … 3.79GB          0 ( 0%)        0%
  cpu_cycles         3.52T  ±  106G     3.46T  … 3.65T           0 ( 0%)        0%
  instructions       2.67T  ± 84.4G     2.62T  … 2.77T           0 ( 0%)        0%
  cache_references    161G  ± 3.52G      159G  …  165G           0 ( 0%)        0%
  cache_misses       20.0G  ±  336M     19.8G  … 20.4G           0 ( 0%)        0%
  branch_misses      3.88G  ±  116M     3.81G  … 4.01G           0 ( 0%)        0%
Benchmark 2 (3 runs): branch/stage4/bin/zig build-exe ...
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           469s  ± 3.71s      466s  …  473s           0 ( 0%)        💩+ 10.9% ±  2.3%
  peak_rss           3.56GB ± 36.3MB    3.54GB … 3.60GB          0 ( 0%)          -  1.0% ±  7.6%
  cpu_cycles         3.78T  ± 21.7G     3.76T  … 3.80T           0 ( 0%)        💩+  7.2% ±  4.9%
  instructions       2.78T  ± 17.3G     2.77T  … 2.80T           0 ( 0%)          +  4.1% ±  5.2%
  cache_references    170G  ±  602M      169G  …  170G           0 ( 0%)        💩+  5.2% ±  3.6%
  cache_misses       20.4G  ± 50.1M     20.3G  … 20.4G           0 ( 0%)          +  1.6% ±  2.7%
  branch_misses      4.01G  ± 24.3M     3.99G  … 4.04G           0 ( 0%)          +  3.4% ±  4.9%
```

### llvm backend Debug vs x86_64 backend Debug, building a subset of the Debug Zig compiler from this branch

```
Benchmark 1 (3 runs): master-llvm/bin/zig build-exe ...
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          19.6s  ± 85.8ms    19.5s  … 19.7s           0 ( 0%)        0%
  peak_rss            798MB ± 4.74MB     795MB …  803MB          0 ( 0%)        0%
  cpu_cycles          220G  ±  285M      219G  …  220G           0 ( 0%)        0%
  instructions        485G  ± 5.71M      485G  …  485G           0 ( 0%)        0%
  cache_references   20.2G  ± 12.0M     20.2G  … 20.2G           0 ( 0%)        0%
  cache_misses       1.05G  ± 8.68M     1.04G  … 1.06G           0 ( 0%)        0%
  branch_misses       416M  ±  843K      415M  …  416M           0 ( 0%)        0%
Benchmark 2 (3 runs): master-x86_64/bin/zig build-exe ...
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          58.5s  ±  190ms    58.3s  … 58.7s           0 ( 0%)        💩+198.8% ±  1.7%
  peak_rss            897MB ± 10.6MB     886MB …  907MB          0 ( 0%)        💩+ 12.5% ±  2.3%
  cpu_cycles          645G  ±  921M      644G  …  646G           0 ( 0%)        💩+193.5% ±  0.7%
  instructions        559G  ± 3.89M      559G  …  559G           0 ( 0%)        💩+ 15.2% ±  0.0%
  cache_references   59.8G  ± 28.4M     59.8G  … 59.8G           0 ( 0%)        💩+196.1% ±  0.2%
  cache_misses       3.63G  ± 36.6M     3.60G  … 3.67G           0 ( 0%)        💩+245.3% ±  5.7%
  branch_misses       665M  ± 5.26M      659M  …  669M           0 ( 0%)        💩+ 60.1% ±  2.1%
```
```
Benchmark 1 (3 runs): branch-llvm/llvm/bin/zig build-exe ...
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          28.2s  ± 1.17s     27.5s  … 29.6s           0 ( 0%)        0%
  peak_rss            960MB ±  199MB     841MB … 1.19GB          0 ( 0%)        0%
  cpu_cycles          314G  ± 36.3G      293G  …  356G           0 ( 0%)        0%
  instructions        611G  ± 74.2G      568G  …  696G           0 ( 0%)        0%
  cache_references   26.9G  ± 2.43G     25.5G  … 29.8G           0 ( 0%)        0%
  cache_misses       1.60G  ±  127M     1.52G  … 1.75G           0 ( 0%)        0%
  branch_misses       518M  ±  103M      458M  …  637M           0 ( 0%)        0%
Benchmark 2 (3 runs): branch-x86_64/debug/bin/zig build-exe ...
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          70.9s  ± 2.95s     69.0s  … 74.3s           0 ( 0%)        💩+151.1% ± 18.0%
  peak_rss            992MB ±  210MB     866MB … 1.23GB          0 ( 0%)          +  3.3% ± 48.3%
  cpu_cycles          775G  ± 91.2G      721G  …  880G           0 ( 0%)        💩+146.6% ± 50.1%
  instructions        655G  ± 94.2G      601G  …  764G           0 ( 0%)          +  7.3% ± 31.5%
  cache_references   69.1G  ± 5.49G     65.6G  … 75.4G           0 ( 0%)        💩+156.4% ± 35.7%
  cache_misses       3.85G  ±  195M     3.74G  … 4.08G           0 ( 0%)        💩+141.0% ± 23.4%
  branch_misses       991M  ±  144M      903M  … 1.16G           0 ( 0%)        💩+ 91.4% ± 55.0%
```