@andrewrk 2025-09-06
lots of this triggered by test-cases:
```
/home/andy/dev/zig/lib/std/Io/Writer.zig:342:5: error: unimplemented ptr_slice_len_ptr
pub fn defaultRebase(w: *Writer, preserve: usize, minimum_len: usize) Error!void {
~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
```
from src/codegen/aarch64/Select.zig line 964