@andrewrk 2025-09-06
Blocking on #25172, which also affects master branch.