@jacobly0 2025-09-08
I guess an argument in favor of not allowing them is that the index operand is dead in a way that is not reflected in `Air.Liveness`, which affects the output size of the llvm and c backends which are not likely to be rewritten to not use `Air.Liveness` anytime soon.