@andrewrk 2025-09-09
Good idea. Yeah it was a pretty trivial fix actually. Seems to be just one missing check inside `analyzePtrArithmetic`.